requirements-test.txt
pytest
pytest-django
pytest-xdist
pytest-cov
factory-boy
